增减了一个工具类，用来查找外键约束

+++ src/tmall/servlet/ProductServlet.java
@@ -1,17 +1,18 @@
 package tmall.servlet;
 
-import tmall.bean.Category;
-import tmall.bean.Product;
-import tmall.bean.Property;
-import tmall.bean.PropertyValue;
+import tmall.bean.*;
+import tmall.dao.ProductImageDAO;
+import tmall.util.ForeignKey;
 import tmall.util.Page;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.File;
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
 
 public class ProductServlet extends BaseBackServlet {
 
@@ -44,8 +45,82 @@ public class ProductServlet extends BaseBackServlet {
     public String delete(HttpServletRequest request, HttpServletResponse response, Page page) {
         int id = Integer.parseInt(request.getParameter("id"));
         Product product = productDAO.get(id);
-        productDAO.delete(id);
-        return "@admin_product_list?cid="+product.getCategory().getId();
+
+        /*
+        //因为存在外间约束，所以根据product,找出所有对应的图片删除
+        List<ProductImage> pisSingle = productImageDAO.list(product, ProductImageDAO.type_single);
+        List<ProductImage> pisDetail=productImageDAO.list(product, ProductImageDAO.type_detail);
+        //删除数据库中图片的信息
+        for (ProductImage piSingle : pisSingle) {
+            productImageDAO.delete(piSingle.getId());
+        }
+        for (ProductImage piDetail : pisDetail) {
+            productImageDAO.delete(piDetail.getId());
+        }
+        //删除对应的图片文件
+        String imageFolder_detail = request.getSession().getServletContext().getRealPath("img/Detail");
+        String imageFolder_single = request.getSession().getServletContext().getRealPath("img/productSingle");
+        String imageFolder_small = request.getSession().getServletContext().getRealPath("img/productSingle_small");
+        String imageFolder_middle = request.getSession().getServletContext().getRealPath("img/productSingle_middle");
+
+
+        File imageFile_detail = new File(imageFolder_detail, id + ".jpg");
+        File imageFile_single = new File(imageFolder_single, id + ".jpg");
+        File imageFile_small = new File(imageFolder_small, id + ".jpg");
+        File imageFile_middle = new File(imageFolder_middle, id + ".jpg");
+
+        imageFile_detail.delete();
+        imageFile_single.delete();
+        imageFile_small.delete();
+        imageFile_middle.delete();
+        */
+
+        Map<String,List> foreignKey = ForeignKey.mapForeignKeys(product);
+
+        List<ProductImage> pis = foreignKey.get("pis");
+        List<PropertyValue> ptvs = foreignKey.get("ptvs");
+        List<OrderItem> ois = foreignKey.get("ois");
+        List<Review> rvs = foreignKey.get("rvs");
+
+        //如果没有关联的外键就可以删除产品
+        if (0 == pis.size() && 0 == ois.size() && 0 == rvs.size()) {
+            //因为属性值是自动初始化的，且始终存在即使值为null，所以删除产品时应自动关联删除其属性值
+            if (0 < ptvs.size()) {
+                for (PropertyValue ptv : ptvs) {
+                    propertyValueDAO.delete(ptv.getId());//删除对应的属性值
+                }
+            }
+            productDAO.delete(id);
+            return "@admin_product_list?cid="+product.getCategory().getId();
+        }
+
+        //外键信息
+
+        String fkInfo = "产品(name)：" + product.getName() + "无法删除，原因：外键约束\n\n";
+        if(0<pis.size()){
+            String piInfo="产品图片(id)：";
+            for (ProductImage pi : pis) {
+                piInfo+=(pi.getId()+" ");
+            }
+            fkInfo += piInfo + "\n";
+        }
+
+        if (0 < ois.size()) {
+            String oiInfo="订单项(id)：";
+            for (OrderItem oi : ois) {
+                oiInfo = oi.getId() + " ";
+            }
+            fkInfo += oiInfo + "\n";
+        }
+        if (0 < rvs.size()) {
+            String rvInfo="产品评论(id)：";
+            for (Review rv : rvs) {
+                rvInfo = rv.getId()+" ";
+            }
+            fkInfo += rvInfo + "\n";
+        }
+
+        return "%"+fkInfo;
     }
 
     @Override
